+++ migrations/V0040__media_attachment__add_file_size.sql
@@ -1,0 +1,1 @@
+ALTER TABLE media_attachment ADD COLUMN file_size INTEGER;