+++ src/activitypub/handlers/create.rs
@@ -215,7 +215,7 @@ pub async fn handle_note(
             };
             let attachment_url = attachment.url
                 .ok_or(ValidationError("attachment URL is missing"))?;
-            let (file_name, maybe_media_type) = fetch_file(
+            let (file_name, file_size, maybe_media_type) = fetch_file(
                 instance,
                 &attachment_url,
                 attachment.media_type.as_deref(),
@@ -227,18 +227,19 @@ pub async fn handle_note(
                     ValidationError("failed to fetch attachment")
                 })?;
             log::info!("downloaded attachment {}", attachment_url);
-            downloaded.push((file_name, maybe_media_type));
+            downloaded.push((file_name, file_size, maybe_media_type));
             // Stop downloading if limit is reached
             if downloaded.len() >= ATTACHMENTS_MAX_NUM {
                 log::warn!("too many attachments");
                 break;
             };
         };
-        for (file_name, maybe_media_type) in downloaded {
+        for (file_name, file_size, maybe_media_type) in downloaded {
             let db_attachment = create_attachment(
                 db_client,
                 &author.id,
                 file_name,
+                file_size,
                 maybe_media_type,
             ).await?;
             attachments.push(db_attachment.id);
@@ -411,7 +412,7 @@ pub async fn handle_note(
                     Err(DatabaseError::NotFound("emoji")) => None,
                     Err(other_error) => return Err(other_error.into()),
                 };
-                let (file_name, maybe_media_type) = match fetch_file(
+                let (file_name, file_size, maybe_media_type) = match fetch_file(
                     instance,
                     &tag.icon.url,
                     tag.icon.media_type.as_deref(),
@@ -437,7 +438,7 @@ pub async fn handle_note(
                     },
                 };
                 log::info!("downloaded emoji {}", tag.icon.url);
-                let image = EmojiImage { file_name, media_type };
+                let image = EmojiImage { file_name, file_size, media_type };
                 let emoji = if let Some(emoji_id) = maybe_emoji_id {
                     update_emoji(
                         db_client,