+++ src/mastodon_api/media/views.rs
@@ -19,7 +19,7 @@ async fn create_attachment_view(
 ) -> Result<HttpResponse, HttpError> {
     let db_client = &**get_database_client(&db_pool).await?;
     let current_user = get_current_user(db_client, auth.token()).await?;
-    let (file_name, media_type) = save_b64_file(
+    let (file_name, file_size, media_type) = save_b64_file(
         &attachment_data.file,
         attachment_data.media_type.clone(),
         &config.media_dir(),
@@ -29,6 +29,7 @@ async fn create_attachment_view(
         db_client,
         &current_user.id,
         file_name,
+        file_size,
         Some(media_type),
     ).await?;
     let attachment = Attachment::from_db(