+++ src/models/emojis/types.rs
@@ -4,10 +4,15 @@ use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use crate::database::json_macro::{json_from_sql, json_to_sql};
+use super::validators::EMOJI_MAX_SIZE;
+
+fn default_emoji_file_size() -> usize { EMOJI_MAX_SIZE }
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct EmojiImage {
     pub file_name: String,
+    #[serde(default = "default_emoji_file_size")]
+    pub file_size: usize,
     pub media_type: String,
 }