+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Save sizes of media attachments and other files to database.
+
 ### Security
 
 - Validate emoji name before saving.